finished library to connect to database

+++ SI2App/SI2App/Model/Conference.cs
@@ -14,7 +14,7 @@
 
         public string Acronym { get; set; }
 
-        public int? Grade { get; set; }
+        public float? Grade { get; set; }
 
         public DateTime? SubmissionDate { get; set; }
